requirements.txt
chandra-ocr==0.1.0
vllm==0.6.1
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pillow==10.1.0
pydantic==2.5.0
python-dotenv==1.0.0
torch==2.1.1
transformers==4.36.0
